fix(formatting): cover user geo edge cases
Render user geo in the unsupported-event path and route all geo summary formatting through a shared helper so issue and search output stay consistent.

Update the unsupported event regression to assert the geo output that the formatter now preserves.

Refs #893
