+++ packages/mcp-core/src/tools/search-events/formatters.ts
@@ -1,4 +1,5 @@
 import type { SentryApiService } from "../../api-client";
+import { formatUserGeoSummary } from "../../internal/user-formatting";
 import { logInfo } from "../../telem/logging";
 import {
   type FlexibleEventData,
@@ -6,31 +7,6 @@ import {
   getStringValue,
   isAggregateQuery,
 } from "./utils";
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-function formatUserGeoValue(value: unknown): string | null {
-  if (!isRecord(value)) {
-    return null;
-  }
-
-  const parts = [
-    value.country_code,
-    value.city,
-    value.region,
-    value.country_name,
-  ].filter(
-    (part): part is string => typeof part === "string" && part.length > 0,
-  );
-
-  if (parts.length === 0) {
-    return null;
-  }
-
-  return Array.from(new Set(parts)).join(", ");
-}
 
 /**
  * Format an explanation for how a natural language query was translated
@@ -164,8 +140,10 @@ export function formatErrorResults(params: FormatEventResultsParams): string {
         ) {
           output += `**${key}**: ${formatEventValue(value)}\n`;
 
-          if (key === "user" && isRecord(value)) {
-            const geoSummary = formatUserGeoValue(value.geo);
+          if (key === "user" && typeof value === "object" && value !== null) {
+            const geoSummary = formatUserGeoSummary(
+              (value as Record<string, unknown>).geo,
+            );
             if (geoSummary) {
               output += `**user.geo**: ${geoSummary}\n`;
             }
